backend/requirements.txt
# KVM RPA 系统依赖包
# Python 版本要求: >=3.9

# 注意事项:
# 1. Sophon硬件环境需要单独安装 sophon-sail (由Sophon SDK提供)
# 2. KVM客户端依赖已通过源代码集成在 src/python_client 目录
# 3. 开发环境可使用模拟模式,不需要实际硬件

# 视频采集与图像处理
opencv-python==4.8.1.78
numpy==1.24.3

# KVM客户端依赖
pydes>=2.0.1

# Sophon SAIL (需要在Sophon硬件环境中安装)
# sophon-sail  # 由Sophon SDK提供,不在pip中

# Web 框架与 API
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6

# 配置管理
pyyaml==6.0.1
python-dotenv==1.0.0

# 日志与监控
loguru==0.7.2
prometheus-client==0.19.0

# HTTP 客户端
requests==2.31.0
httpx==0.25.1

# 工具库
python-dateutil==2.8.2
pydantic==2.5.0

